+++ resources/js/filament/filament-rich-editor-extender/Youtube.js
@@ -1,0 +1,5 @@
+import Youtube from '@tiptap/extension-youtube'
+
+export default Youtube.configure({
+    nocookie: true,
+})