+++ resources/lang/pt_BR/youtube.php
@@ -1,0 +1,23 @@
+<?php
+
+return [
+
+    'tool' => [
+        'label' => 'YouTube',
+    ],
+
+    'action' => [
+        'label' => 'Inserir vídeo do YouTube',
+
+        'modal' => [
+            'heading' => 'Inserir vídeo do YouTube',
+
+            'form' => [
+                'url' => [
+                    'label' => 'URL',
+                ],
+            ],
+        ],
+    ],
+
+];