+++ src/FilamentRichEditorExtenderServiceProvider.php
@@ -2,6 +2,10 @@
 
 namespace MateusLecchi\FilamentRichEditorExtender;
 
+use Filament\Forms\Components\RichEditor;
+use Filament\Support\Assets\Js;
+use Filament\Support\Facades\FilamentAsset;
+use MateusLecchi\FilamentRichEditorExtender\Plugins\YoutubePlugin;
 use Spatie\LaravelPackageTools\Commands\InstallCommand;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
@@ -17,6 +21,7 @@ class FilamentRichEditorExtenderServiceProvider extends PackageServiceProvider
          */
         $package
             ->name('filament-rich-editor-extender')
+            ->hasTranslations()
             ->hasInstallCommand(function (InstallCommand $command): void {
                 $command->askToStarRepoOnGitHub('mateuslecchi/filament-rich-editor-extender');
             });
@@ -24,16 +29,14 @@ class FilamentRichEditorExtenderServiceProvider extends PackageServiceProvider
 
     public function bootingPackage(): void
     {
-        // Register your extensions here as they are added, e.g.:
-        //
-        // FilamentAsset::register([
-        //     Js::make('filament-rich-editor-extender/your-extension', __DIR__.'/../resources/js/dist/filament/filament-rich-editor-extender/YourExtension.js')->loadedOnRequest(),
-        // ]);
-        //
-        // RichEditor::configureUsing(function (RichEditor $richEditor) {
-        //     $richEditor->plugins([
-        //         YourPlugin::make(),
-        //     ]);
-        // });
+        FilamentAsset::register([
+            Js::make('filament-rich-editor-extender/youtube', __DIR__.'/../resources/js/dist/filament/filament-rich-editor-extender/Youtube.js')->loadedOnRequest(),
+        ]);
+
+        RichEditor::configureUsing(function (RichEditor $richEditor): void {
+            $richEditor->plugins([
+                YoutubePlugin::make(),
+            ]);
+        });
     }
 }